add modusdb inference history for local dev
