+++ runtime/.gitignore
@@ -3,3 +3,5 @@ runtime
 
 node_modules
 dist
+
+data
