+++ runtime/httpserver/server.go
@@ -23,6 +23,7 @@ import (
 	"github.com/hypermodeinc/modus/lib/manifest"
 	"github.com/hypermodeinc/modus/runtime/app"
 	"github.com/hypermodeinc/modus/runtime/config"
+	"github.com/hypermodeinc/modus/runtime/db"
 	"github.com/hypermodeinc/modus/runtime/explorer"
 	"github.com/hypermodeinc/modus/runtime/graphql"
 	"github.com/hypermodeinc/modus/runtime/logger"
@@ -137,6 +138,8 @@ func GetMainHandler(options ...func(map[string]http.Handler)) http.Handler {
 	if config.IsDevEnvironment() {
 		defaultRoutes["/explorer/"] = explorer.ExplorerHandler
 		defaultRoutes["/"] = http.RedirectHandler("/explorer/", http.StatusSeeOther)
+		defaultRoutes["/inferences.json"] = InferenceHistoryHandler
+		defaultRoutes["/plugins.json"] = PluginHandler
 	}
 
 	for _, opt := range options {
@@ -256,4 +259,25 @@ func isIPv6Available() bool {
 	}
 	defer conn.Close()
 	return true
+}
+
+var InferenceHistoryHandler = http.HandlerFunc(inferenceHistoryHandler)
+var PluginHandler = http.HandlerFunc(pluginHandler)
+
+func inferenceHistoryHandler(w http.ResponseWriter, r *http.Request) {
+	resp, err := db.QueryInferences()
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	fmt.Fprint(w, resp)
+}
+
+func pluginHandler(w http.ResponseWriter, r *http.Request) {
+	resp, err := db.QueryPlugins()
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	fmt.Fprint(w, resp)
 }
