+++ runtime/services/services.go
@@ -54,6 +54,7 @@ func Start(ctx context.Context) context.Context {
 	secrets.Initialize(ctx)
 	storage.Initialize(ctx)
 	db.Initialize(ctx)
+	db.InitModusDb(ctx)
 	collections.Initialize(ctx)
 	manifestdata.MonitorManifestFile(ctx)
 	envfiles.MonitorEnvFiles(ctx)
@@ -81,4 +82,5 @@ func Stop(ctx context.Context) {
 	neo4jclient.CloseDrivers(ctx)
 	logger.Close()
 	db.Stop(ctx)
+	db.CloseModusDb(ctx)
 }
